+++ examples/counter.rs
@@ -1,5 +1,5 @@
 use pretty_env_logger;
-use snurr::Process;
+use snurr::ProcessBuilder;
 use std::sync::atomic::{AtomicU32, Ordering::Relaxed};
 
 #[derive(Debug, Default)]
@@ -9,7 +9,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     pretty_env_logger::init();
 
     // Create process from BPMN file
-    let bpmn = Process::<Counter>::new("examples/counter.bpmn")?
+    let bpmn = ProcessBuilder::<Counter>::new("examples/counter.bpmn")?
         .task("Count 1", |input| {
             input.0.fetch_add(1, Relaxed);
             Default::default()